Made revolver scripting. Made it super easy to add new units.

+++ src/bibble/data/data_types/effect_unit.rs
@@ -8,7 +8,7 @@ pub enum EffectToUnit{
 }
 
 pub struct EffectToUnitDamage{
-    pub amount: u32,
+    pub amount: f32,
 }
 
 pub struct EffectToUnitAddWeapon{
